+++ src/config/index.tsx
@@ -15,6 +15,6 @@ const config = {
   },
 };
 
-export type EndpointType = 'getPokemons';
+export type EndpointType = keyof typeof config.client.endpoint;
 
 export default config;
