+++ src/utils/request.ts
@@ -2,13 +2,10 @@ import Url from 'url';
 import getUrlWithParamsConfig from './getUrlWithParamsConfig';
 import { EndpointType } from '../config';
 
-async function req(endpoint: EndpointType, query?: object) {
+async function req<T>(endpoint: EndpointType, query?: object): Promise<T> {
   const uri = Url.format(getUrlWithParamsConfig(endpoint, query));
-  console.log('req -> uri', uri);
-
   const response = await fetch(uri);
   const result = await response.json();
-
   return result;
 }
